debt: remove repo-sandbox
